refactored the code to put flash version checking in separate file
