+++ llmfoundry/__init__.py
@@ -9,6 +9,9 @@ try:
     # gated import otherwise.
     import transformers
 
+    from llmfoundry.models.utils.flash_attn_checker import is_flash_v1_installed
+    if is_flash_v1_installed():
+        transformers.utils.is_flash_attn_available = lambda: False
     from llmfoundry import optim, utils
     from llmfoundry.data import (ConcatTokensDataset,
                                  MixtureOfDenoisersCollator, NoConcatDataset,
@@ -19,8 +22,8 @@ try:
                                       ComposerHFT5)
     from llmfoundry.models.layers.attention import (
         MultiheadAttention, attn_bias_shape, build_alibi_bias, build_attn_bias,
-        flash_attn_fn, is_flash_v1_installed,
-        scaled_multihead_dot_product_attention, triton_flash_attn_fn)
+        flash_attn_fn, scaled_multihead_dot_product_attention,
+        triton_flash_attn_fn)
     from llmfoundry.models.layers.blocks import MPTBlock
     from llmfoundry.models.layers.ffn import (FFN_CLASS_REGISTRY, MPTMLP,
                                               build_ffn)
@@ -29,8 +32,6 @@ try:
                                        MPTForCausalLM, MPTModel,
                                        MPTPreTrainedModel)
     from llmfoundry.tokenizers import TiktokenTokenizerWrapper
-    if is_flash_v1_installed():
-        transformers.utils.is_flash_attn_available = lambda: False
 
 except ImportError as e:
     try:
